requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py
asyncio
async_timeout
PyNaCl
aiohttp
requests
pymongo
pillow
chatterbot
nltk
psutil
youtube_dl==2019.1.24
websockets
pyfiglet
pytz